+++ src/main/java/net/tridentsdk/packets/play/in/PacketPlayInPlayerClickWindow.java
@@ -115,7 +115,7 @@ public class PacketPlayInPlayerClickWindow extends InPacket {
     @Override
     public void handleReceived(ClientConnection connection) {
         Window window = TridentServer.getInstance().getWindow(this.windowId);
-        PlayerClickItemEvent clickEvent = new PlayerClickItemEvent(window, this.clickedSlot, this.actionNumber);
+        PlayerClickItemEvent clickEvent = new PlayerClickItemEvent(window, this.clickedSlot, (int) this.actionNumber);
 
         TridentServer.getInstance().getEventManager().call(clickEvent);
